+++ test/knxip_tests/tunnelling_request_test.py
@@ -31,7 +31,7 @@ class TestKNXIPTunnellingRequest:
             payload=GroupValueWrite(DPTBinary(1)),
         )
 
-        cemi = CEMIFrame(xknx, code=CEMIMessageCode.L_DATA_REQ)
+        cemi = CEMIFrame(code=CEMIMessageCode.L_DATA_REQ)
         cemi.telegram = Telegram(
             destination_address=GroupAddress("9/0/8"),
             payload=GroupValueWrite(DPTBinary(1)),