+++ xknx/io/routing.py
@@ -78,7 +78,6 @@ class Routing(Interface):
     async def send_telegram(self, telegram: "Telegram") -> None:
         """Send Telegram to routing connected device."""
         cemi = CEMIFrame.init_from_telegram(
-            self.xknx,
             telegram=telegram,
             code=CEMIMessageCode.L_DATA_IND,
             src_addr=self.xknx.own_address,