+++ changelog.md
@@ -5,6 +5,7 @@
 ### Internal
 
 - Move `levels` instance attribute form `GroupAddress` to `address_format` class variable
+- Remove xknx form CEMIFrame class
 
 ## 0.20.3 Threading fixes 2022-04-15